project/requirements.txt
flask==2.0.1
pyserial==3.5
werkzeug==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
pytest==6.2.5
requests==2.26.0